flask-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
requests==2.31.0
PyPDF2==3.0.1
python-dotenv==1.0.0